Remove FilterContent
