+++ frontend/src/metabase/querying/drills/utils/filter-drill.tsx
@@ -1,4 +1,4 @@
-import { FilterPickerBody } from "metabase/querying/filters/components/FilterPicker/FilterPickerBody";
+import { FilterPickerBody } from "metabase/querying/filters";
 import type { ClickActionPopoverProps } from "metabase/visualizations/types";
 import * as Lib from "metabase-lib";
 import type Question from "metabase-lib/v1/Question";
