+++ frontend/src/metabase/querying/filters/components/FilterModal/DefaultFilterEditor/DefaultFilterEditor.tsx
@@ -1,12 +1,12 @@
 import { useMemo } from "react";
 
 import { getColumnIcon } from "metabase/common/utils/columns";
-import { Checkbox, Grid, Group } from "metabase/ui";
-
 import {
   type OperatorOption,
   useDefaultFilter,
-} from "../../../hooks/use-default-filter";
+} from "metabase/querying/filters/hooks/use-default-filter";
+import { Checkbox, Grid, Group } from "metabase/ui";
+
 import { FilterTitle, HoverParent } from "../FilterTitle";
 import type { FilterEditorProps } from "../types";
