+++ frontend/src/metabase/querying/filters/components/FilterModal/FilterModalBody/SegmentFilterItem/utils.ts
@@ -1,6 +1,5 @@
+import type { SegmentItem } from "metabase/querying/filters/hooks/use-filter-modal";
 import * as Lib from "metabase-lib";
-
-import type { SegmentItem } from "../types";
 
 export function addSegmentFilters(
   query: Lib.Query,
