+++ frontend/src/metabase/querying/filters/components/FilterModal/SegmentFilterEditor/SegmentFilterEditor.tsx
@@ -1,8 +1,7 @@
 import { t } from "ttag";
 
+import type { SegmentItem } from "metabase/querying/filters/hooks/use-filter-modal";
 import { Flex, Grid, Icon, MultiSelect, Text } from "metabase/ui";
-
-import type { SegmentItem } from "../types";
 
 interface SegmentFilterEditorProps {
   segmentItems: SegmentItem[];
