+++ frontend/src/metabase/querying/filters/components/FilterModal/types.ts
@@ -1,26 +1,4 @@
-import type { IconName } from "metabase/ui";
 import type * as Lib from "metabase-lib";
-
-export interface ColumnItem {
-  column: Lib.ColumnMetadata;
-  displayName: string;
-  stageIndex: number;
-}
-
-export interface SegmentItem {
-  segment: Lib.SegmentMetadata;
-  displayName: string;
-  stageIndex: number;
-  filterPositions: number[];
-}
-
-export interface GroupItem {
-  key: string;
-  displayName: string;
-  icon: IconName;
-  columnItems: ColumnItem[];
-  segmentItems: SegmentItem[];
-}
 
 export interface FilterEditorProps {
   query: Lib.Query;
