+++ frontend/src/metabase/querying/filters/hooks/use-filter-modal/index.ts
@@ -1,1 +1,2 @@
 export * from "./use-filter-modal";
+export * from "./types";
