+++ frontend/src/metabase/querying/filters/hooks/use-filter-modal/utils/index.ts
@@ -1,3 +1,2 @@
 export * from "./filters";
-export * from "./modal";
 export * from "./search";
